Federalized robot config.
