+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/subsystems/Claw.kt
@@ -3,6 +3,7 @@ package org.firstinspires.ftc.teamcode.subsystems
 import android.graphics.Color
 import android.util.Log
 import com.acmerobotics.dashboard.canvas.Canvas
+import com.acmerobotics.dashboard.config.Config
 import com.acmerobotics.roadrunner.Pose2d
 import com.acmerobotics.roadrunner.TimeProfile
 import com.acmerobotics.roadrunner.Vector2d
@@ -14,19 +15,13 @@ import com.qualcomm.robotcore.hardware.NormalizedColorSensor
 import com.qualcomm.robotcore.util.ElapsedTime
 import com.qualcomm.robotcore.util.Range.clip
 import org.firstinspires.ftc.robotcore.external.Telemetry
-import org.firstinspires.ftc.teamcode.RobotConfig
-import org.firstinspires.ftc.teamcode.RobotConfig.clawClosedPosition
-import org.firstinspires.ftc.teamcode.RobotConfig.clawOpenedPosition
-import org.firstinspires.ftc.teamcode.RobotConfig.clawPartiallyOpenedPosition
-import org.firstinspires.ftc.teamcode.RobotConfig.clawServoName
-import org.firstinspires.ftc.teamcode.RobotConfig.colorGain
-import org.firstinspires.ftc.teamcode.RobotConfig.colorSensorName
-import org.firstinspires.ftc.teamcode.RobotConfig.valueThreshold
+import org.firstinspires.ftc.teamcode.FieldConfig
 
 /**
  * Claw subsystem consists of a servo, potentiometer and a color sensor.
  * @param hwMap HardwareMap.
  */
+@Config
 class Claw(hwMap: HardwareMap, startingPosition: Double = clawClosedPosition) : SubsystemBase() {
 
     private val servo = SimpleServo(hwMap, clawServoName, 0.0, 360.0)
@@ -39,7 +34,7 @@ class Claw(hwMap: HardwareMap, startingPosition: Double = clawClosedPosition) : 
         set(position) {
             field = clip(position, 0.0, 1.0)
             timer.reset()
-            motionProfile = TimeProfile(constantProfile(position - getPositionEstimate(), 0.0, RobotConfig.clawMaxVel, -RobotConfig.clawMaxAccel, RobotConfig.clawMaxAccel).baseProfile)
+            motionProfile = TimeProfile(constantProfile(position - getPositionEstimate(), 0.0, clawMaxVel, -clawMaxAccel, clawMaxAccel).baseProfile)
             Log.i("Claw desired position", setpoint.toString())
         }
 
@@ -110,7 +105,30 @@ class Claw(hwMap: HardwareMap, startingPosition: Double = clawClosedPosition) : 
                 clawOffset
             )
         )
-        canvas.fillCircle(x, y, RobotConfig.coneDiameter / 2.0 * 1.2)
+        canvas.fillCircle(x, y, FieldConfig.coneDiameter / 2.0 * 1.2)
+    }
+
+    companion object {
+        const val clawServoName = "claw"
+        const val colorSensorName = "color"
+
+        @JvmField
+        var clawClosedPosition = 0.85
+        @JvmField
+        var clawOpenedPosition = 0.73
+        @JvmField
+        var clawPartiallyOpenedPosition = 0.80
+
+
+        @JvmField
+        var clawMaxVel = 5.0
+        @JvmField
+        var clawMaxAccel = 5.0
+
+        @JvmField
+        var colorGain = 20.0
+        @JvmField
+        var valueThreshold = 0.15
     }
 
 }
