+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/teleops/tuning/LiftDistancePerPulseTuner.kt
@@ -2,14 +2,10 @@ package org.firstinspires.ftc.teamcode.teleops.tuning
 
 import com.acmerobotics.dashboard.FtcDashboard
 import com.acmerobotics.dashboard.telemetry.MultipleTelemetry
-import com.qualcomm.robotcore.eventloop.opmode.Disabled
 import com.qualcomm.robotcore.eventloop.opmode.OpMode
 import com.qualcomm.robotcore.eventloop.opmode.TeleOp
-import org.firstinspires.ftc.teamcode.RobotConfig
 import org.firstinspires.ftc.teamcode.subsystems.Lift
 import org.firstinspires.ftc.teamcode.subsystems.VoltageSensor
-import java.lang.Math.toRadians
-import kotlin.math.sin
 
 @TeleOp
 class LiftDistancePerPulseTuner() : OpMode() {
