+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/teleops/tuning/PassthroughTuner.kt
@@ -3,8 +3,11 @@ package org.firstinspires.ftc.teamcode.teleops.tuning
 import com.qualcomm.robotcore.eventloop.opmode.Disabled
 import com.qualcomm.robotcore.eventloop.opmode.OpMode
 import com.qualcomm.robotcore.eventloop.opmode.TeleOp
-import org.firstinspires.ftc.teamcode.RobotConfig
+import org.firstinspires.ftc.teamcode.FieldConfig
 import org.firstinspires.ftc.teamcode.subsystems.Passthrough
+import org.firstinspires.ftc.teamcode.subsystems.Passthrough.Companion.passthroughDepositAngle
+import org.firstinspires.ftc.teamcode.subsystems.Passthrough.Companion.passthroughJunctionAngle
+import org.firstinspires.ftc.teamcode.subsystems.Passthrough.Companion.passthroughPickUpAngle
 
 @TeleOp
 @Disabled
@@ -34,9 +37,9 @@ class PassthroughTuner() : OpMode() {
             passthrough.setPosition(0.0)
         }
 
-        telemetry.addData("Time to deposit", passthrough.timeToTarget(RobotConfig.passthroughDepositAngle))
-        telemetry.addData("Time to junction deposit", passthrough.timeToTarget(RobotConfig.passthroughJunctionAngle))
-        telemetry.addData("Time to pick up", passthrough.timeToTarget(RobotConfig.passthroughPickUpAngle))
+        telemetry.addData("Time to deposit", passthrough.timeToTarget(passthroughDepositAngle))
+        telemetry.addData("Time to junction deposit", passthrough.timeToTarget(passthroughJunctionAngle))
+        telemetry.addData("Time to pick up", passthrough.timeToTarget(passthroughPickUpAngle))
         passthrough.fetchTelemetry(telemetry)
         telemetry.update()
     }
