+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/vision/modules/UndistortLens.kt
@@ -1,6 +1,7 @@
 package org.firstinspires.ftc.teamcode.vision.modules
 
-import org.firstinspires.ftc.teamcode.RobotConfig
+import org.firstinspires.ftc.teamcode.FieldConfig
+import org.firstinspires.ftc.teamcode.subsystems.Vision
 import org.firstinspires.ftc.teamcode.vision.modulelib.AbstractPipelineModule
 import org.opencv.calib3d.Calib3d
 import org.opencv.core.Mat
@@ -8,7 +9,7 @@ import org.opencv.core.Mat
 
 class UndistortLens (
     private val inputModule: AbstractPipelineModule<Mat>,
-    private val camera: RobotConfig.CameraData
+    private val camera: Vision.Companion.CameraData
 ) : AbstractPipelineModule<Mat>() {
 
     private lateinit var output: Mat
